Move the waste balance ledger off the admin path
