+++ src/plugins/router.js
@@ -25,12 +25,12 @@ import { summaryLogUploadsReportRoutes } from '#routes/v1/organisations/registra
 import * as reportsRoutes from '#reports/routes/index.js'
 import { reportsUnsubmit } from '#reports/routes/unsubmit.js'
 import { adminMeGet } from '#routes/v1/admin/me/get.js'
-import { accreditationLedgerEventsGet } from '#routes/v1/admin/organisations/registrations/accreditations/ledger-events/get.js'
+import { accreditationWasteBalanceLedgerGet } from '#routes/v1/organisations/registrations/accreditations/waste-balance-ledger/get.js'
 import {
   registrationAccreditationsGet,
   registrationGet
 } from '#routes/v1/organisations/registrations/get.js'
-import { registrationLedgerEventsGet } from '#routes/v1/admin/organisations/registrations/ledger-events/get.js'
+import { registrationWasteBalanceLedgerGet } from '#routes/v1/organisations/registrations/waste-balance-ledger/get.js'
 import { dlqMessagesGet } from '#routes/v1/admin/queues/dlq/messages.get.js'
 import { dlqPurgePost } from '#routes/v1/admin/queues/dlq/purge.post.js'
 
@@ -81,10 +81,10 @@ const router = {
           ...Object.values(coreReportsRoutes),
           reportsUnsubmit,
           adminMeGet,
-          accreditationLedgerEventsGet,
+          accreditationWasteBalanceLedgerGet,
           registrationGet,
           registrationAccreditationsGet,
-          registrationLedgerEventsGet,
+          registrationWasteBalanceLedgerGet,
           dlqMessagesGet,
           dlqPurgePost
         ])
