+++ src/routes/v1/organisations/registrations/waste-balance-ledger/get.js
@@ -1,16 +1,16 @@
 import { StatusCodes } from 'http-status-codes'
 import { SCOPES } from '#common/helpers/auth/constants.js'
 import { readLedger } from '#waste-balances/application/read-ledger.js'
-import { ledgerEventsResponseSchema } from '../ledger-events-response.schema.js'
+import { wasteBalanceLedgerResponseSchema } from '../waste-balance-ledger-response.schema.js'
 
 /** @import { HapiRequest, HapiResponseToolkit } from '#common/hapi-types.js' */
 
-export const registrationLedgerEventsGetPath =
-  '/v1/admin/organisations/{organisationId}/registrations/{registrationId}/waste-balance-events'
+export const registrationWasteBalanceLedgerGetPath =
+  '/v1/organisations/{organisationId}/registrations/{registrationId}/waste-balance-ledger'
 
-export const registrationLedgerEventsGet = {
+export const registrationWasteBalanceLedgerGet = {
   method: 'GET',
-  path: registrationLedgerEventsGetPath,
+  path: registrationWasteBalanceLedgerGetPath,
   options: {
     auth: {
       scope: [
@@ -18,9 +18,9 @@ export const registrationLedgerEventsGet = {
         `+${SCOPES.wasteBalanceLedgerRead}`
       ]
     },
-    tags: ['api', 'admin'],
+    tags: ['api'],
     response: {
-      schema: ledgerEventsResponseSchema
+      schema: wasteBalanceLedgerResponseSchema
     }
   },
   /**
