+++ src/routes/v1/organisations/registrations/waste-balance-ledger/get.test.js
@@ -1,31 +1,29 @@
 import { describe, it, expect, beforeEach } from 'vitest'
 import { StatusCodes } from 'http-status-codes'
 
-import {
-  buildLedgerEvent,
-  buildPrnCreatedEvent
-} from '#waste-balances/repository/ledger-test-data.js'
+import { buildLedgerEvent } from '#waste-balances/repository/ledger-test-data.js'
 import { createTestServer } from '#test/create-test-server.js'
 import { asServiceMaintainer } from '#test/inject-auth.js'
 import { setupAuthContext } from '#vite/helpers/setup-auth-mocking.js'
 import {
-  testLedgerEventsAccess,
-  testLedgerEventsResponseIsChecked
-} from '../../ledger-events-test-helpers.js'
-import { accreditationLedgerEventsGetPath } from './get.js'
+  testWasteBalanceLedgerAccess,
+  testWasteBalanceLedgerResponseIsChecked
+} from '../waste-balance-ledger-test-helpers.js'
+import {
+  registrationWasteBalanceLedgerGet,
+  registrationWasteBalanceLedgerGetPath
+} from './get.js'
 
 /**
  * @param {string} orgId
  * @param {string} regId
- * @param {string} accId
  */
-const makePath = (orgId, regId, accId) =>
-  accreditationLedgerEventsGetPath
+const makePath = (orgId, regId) =>
+  registrationWasteBalanceLedgerGetPath
     .replace('{organisationId}', orgId)
     .replace('{registrationId}', regId)
-    .replace('{accreditationId}', accId)
 
-describe(`GET ${accreditationLedgerEventsGetPath}`, () => {
+describe(`GET ${registrationWasteBalanceLedgerGetPath}`, () => {
   setupAuthContext()
 
   /** @type {import('#test/create-test-server.js').TestServer} */
@@ -38,19 +36,31 @@ describe(`GET ${accreditationLedgerEventsGetPath}`, () => {
     ledgerRepository = server.app.ledgerRepository
   })
 
-  it('returns 200 with events for the ledger', async () => {
+  describe('route metadata', () => {
+    it('exposes the expected method and path', () => {
+      expect(registrationWasteBalanceLedgerGet.method).toBe('GET')
+      expect(registrationWasteBalanceLedgerGet.path).toBe(
+        registrationWasteBalanceLedgerGetPath
+      )
+      expect(registrationWasteBalanceLedgerGetPath).toBe(
+        '/v1/organisations/{organisationId}/registrations/{registrationId}/waste-balance-ledger'
+      )
+    })
+  })
+
+  it('returns 200 with the events of the ledger held with no accreditation', async () => {
     await ledgerRepository.appendEvents([
       buildLedgerEvent({
-        registrationId: 'reg-1',
-        accreditationId: 'acc-1',
         organisationId: 'org-1',
+        registrationId: 'reg-1',
+        accreditationId: null,
         number: 1
       })
     ])
 
     const response = await server.inject({
       method: 'GET',
-      url: makePath('org-1', 'reg-1', 'acc-1'),
+      url: makePath('org-1', 'reg-1'),
       ...asServiceMaintainer()
     })
 
@@ -59,66 +69,68 @@ describe(`GET ${accreditationLedgerEventsGetPath}`, () => {
     expect(result.ledger).toEqual({
       organisationId: 'org-1',
       registrationId: 'reg-1',
-      accreditationId: 'acc-1'
+      accreditationId: null
     })
     expect(result.events).toHaveLength(1)
     expect(result.events[0].number).toBe(1)
   })
 
-  it('gives a PRN event the note it concerns and the note’s tonnage', async () => {
-    await ledgerRepository.appendEvents([
-      buildPrnCreatedEvent({
-        registrationId: 'reg-prn',
-        accreditationId: 'acc-prn',
-        organisationId: 'org-prn',
+  it('gives a submission event its balances, its actor and the log it credits', async () => {
+    const [stored] = await ledgerRepository.appendEvents([
+      buildLedgerEvent({
+        organisationId: 'org-shape',
+        registrationId: 'reg-shape',
+        accreditationId: null,
         number: 1,
-        payload: { prnId: 'prn-1', amount: 50 },
-        openingBalance: { amount: 100, availableAmount: 100 },
-        closingBalance: { amount: 100, availableAmount: 50 }
+        payload: { summaryLogId: 'log-1', creditTotal: 100 },
+        openingBalance: { amount: 0, availableAmount: 0 },
+        closingBalance: { amount: 100, availableAmount: 100 },
+        createdBy: { id: 'user-1', name: 'Test User' }
       })
     ])
 
     const response = await server.inject({
       method: 'GET',
-      url: makePath('org-prn', 'reg-prn', 'acc-prn'),
+      url: makePath('org-shape', 'reg-shape'),
       ...asServiceMaintainer()
     })
 
     expect(response.statusCode).toBe(StatusCodes.OK)
     expect(JSON.parse(response.payload).events[0]).toEqual({
+      id: stored.id,
       number: 1,
-      kind: 'prn-created',
+      kind: 'summary-log-submitted',
       createdAt: '2026-01-15T10:00:00.000Z',
       createdBy: { id: 'user-1', name: 'Test User' },
       balance: {
-        opening: { total: 100, available: 100 },
-        closing: { total: 100, available: 50 }
+        opening: { total: 0, available: 0 },
+        closing: { total: 100, available: 100 }
       },
-      prn: { id: 'prn-1', tonnage: 50 }
+      summaryLog: { id: 'log-1', creditTotal: 100 }
     })
   })
 
   it('returns events ordered by number ascending', async () => {
     await ledgerRepository.appendEvents([
       buildLedgerEvent({
-        registrationId: 'reg-2',
-        accreditationId: 'acc-2',
         organisationId: 'org-2',
+        registrationId: 'reg-2',
+        accreditationId: null,
         number: 1
       })
     ])
     await ledgerRepository.appendEvents([
       buildLedgerEvent({
-        registrationId: 'reg-2',
-        accreditationId: 'acc-2',
         organisationId: 'org-2',
+        registrationId: 'reg-2',
+        accreditationId: null,
         number: 2
       })
     ])
 
     const response = await server.inject({
       method: 'GET',
-      url: makePath('org-2', 'reg-2', 'acc-2'),
+      url: makePath('org-2', 'reg-2'),
       ...asServiceMaintainer()
     })
 
@@ -128,10 +140,30 @@ describe(`GET ${accreditationLedgerEventsGetPath}`, () => {
     expect(result.events.map((event) => event.number)).toEqual([1, 2])
   })
 
-  it('returns no events when the accreditation holds none', async () => {
+  it('returns no events when the registration holds none', async () => {
     const response = await server.inject({
       method: 'GET',
-      url: makePath('org-none', 'reg-none', 'acc-none'),
+      url: makePath('org-none', 'reg-none'),
+      ...asServiceMaintainer()
+    })
+
+    expect(response.statusCode).toBe(StatusCodes.OK)
+    expect(JSON.parse(response.payload).events).toEqual([])
+  })
+
+  it('does not return the events of an accreditation of the same registration', async () => {
+    await ledgerRepository.appendEvents([
+      buildLedgerEvent({
+        organisationId: 'org-3',
+        registrationId: 'reg-3',
+        accreditationId: 'acc-3',
+        number: 1
+      })
+    ])
+
+    const response = await server.inject({
+      method: 'GET',
+      url: makePath('org-3', 'reg-3'),
       ...asServiceMaintainer()
     })
 
@@ -144,14 +176,14 @@ describe(`GET ${accreditationLedgerEventsGetPath}`, () => {
       buildLedgerEvent({
         organisationId: 'org-owner',
         registrationId: 'reg-owned',
-        accreditationId: 'acc-owned',
+        accreditationId: null,
         number: 1
       })
     ])
 
     const response = await server.inject({
       method: 'GET',
-      url: makePath('org-stranger', 'reg-owned', 'acc-owned'),
+      url: makePath('org-stranger', 'reg-owned'),
       ...asServiceMaintainer()
     })
 
@@ -160,16 +192,16 @@ describe(`GET ${accreditationLedgerEventsGetPath}`, () => {
   })
 })
 
-testLedgerEventsAccess({
-  makeUrl: (organisationId) => makePath(organisationId, 'reg-1', 'acc-1')
+testWasteBalanceLedgerAccess({
+  makeUrl: (organisationId) => makePath(organisationId, 'reg-1')
 })
 
-testLedgerEventsResponseIsChecked({
+testWasteBalanceLedgerResponseIsChecked({
   ledgerId: {
     organisationId: 'org-1',
     registrationId: 'reg-1',
-    accreditationId: 'acc-1'
+    accreditationId: null
   },
-  makeUrl: ({ organisationId, registrationId, accreditationId }) =>
-    makePath(organisationId, registrationId, String(accreditationId))
+  makeUrl: ({ organisationId, registrationId }) =>
+    makePath(organisationId, registrationId)
 })
